+++ lib/closure_tree/support.rb
@@ -106,9 +106,7 @@ module ClosureTree
     def with_advisory_lock(&block)
       if options[:with_advisory_lock]
         model_class.with_advisory_lock("closure_tree") do
-          model_class.transaction do
-            yield
-          end
+          transaction { yield }
         end
       else
         yield